# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Team Health Check API — Pulse Framework
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /my-journey:
    get:
      operationId: getMyJourney
      tags: [dashboard]
      summary: Get personal response history and trends
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 90
      responses:
        "200":
          description: Personal journey data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JourneyData"
  /team-summary:
    get:
      operationId: getTeamSummary
      tags: [dashboard]
      summary: Get anonymized team health summary (leads/directors only)
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 90
      responses:
        "200":
          description: Team summary with pillar scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSummaryData"
...
components:
  schemas:
...
    PillarScore:
      type: object
      properties:
        pillar:
          type: string
        score:
          type: number
        trend:
          type: string
          enum: [up, stable, down]
        previousScore:
          type: ["number", "null"]
        favorability:
          type: number
        responseCount:
          type: integer
        respondentCount:
          type: integer
        status:
          type: string
          enum: [green, yellow, red, insufficient]
      required: [pillar, score, trend, previousScore, favorability, responseCount, respondentCount, status]
...
    TrendPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
      required: [date, value]
    JourneyData:
      type: object
      properties:
        totalCheckIns:
          type: integer
        currentStreak:
          type: integer
        pillarTimelines:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/TrendPoint"
        pillarAverages:
          type: array
          items:
            $ref: "#/components/schemas/PillarScore"
        recentReflections:
          type: array
          items:
            $ref: "#/components/schemas/ReflectionEntry"
      required: [totalCheckIns, currentStreak, pillarTimelines, pillarAverages, recentReflections]
    ReflectionEntry:
      type: object
      properties:
        id:
          type: integer
        questionText:
          type: string
        response:
          type: string
        pillar:
          type: string
        date:
          type: string
          format: date-time
      required: [id, questionText, response, pillar, date]
    TeamSummaryData:
      type: object
      properties:
        compositeScore:
          type: number
        participationRate:
          type: number
        respondentCount:
          type: integer
        pillarScores:
          type: array
          items:
            $ref: "#/components/schemas/PillarScore"
        trendChart:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        topStrengths:
          type: array
          items:
            type: string
        areasForGrowth:
          type: array
          items:
            type: string
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/AlertItem"
      required: [compositeScore, participationRate, respondentCount, pillarScores, trendChart, topStrengths, areasForGrowth, alerts]
    AlertItem:
      type: object
      properties:
        pillar:
          type: string
        type:
          type: string
          enum: [decline, critical, polarization]
        message:
          type: string
      required: [pillar, type, message]

    # ─── Design Ops: Projects ──────────────────────────────────────────
...