# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Health Check API — Pulse Framework
 * OpenAPI spec version: 0.2.0
 */

export interface TrendPoint {
  date: string;
  value: number;
}